refactor: 회원가입 페이지 앱라우터로 마이그레이션
