+++ app/layout.tsx
@@ -1,4 +1,5 @@
-import './global.css';
+import QueryProvider from "@/ui/providers/queryProvider";
+import "./global.css";
 
 export default function RootLayout({
   children,
@@ -7,7 +8,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ko">
-      <body>{children}</body>
+      <body>
+        <QueryProvider>{children}</QueryProvider>
+      </body>
     </html>
   );
 }
