+++ app/signup/page.tsx
@@ -1,11 +1,11 @@
-import Image from "next/image";
 import Link from "next/link";
-import LoginForm from "./_components/loginForm";
+import Image from "next/image";
+import SignUpForm from "./_components/signupForm";
 import LogoIcon from "@/public/logo.svg";
 import kakaoIcon from "@/public/Kakao.svg";
 import googleIcon from "@/public/googleIcon.png";
 
-export default function Login() {
+function SignUp() {
   return (
     <div className="flex h-[100vh] flex-col items-center justify-center bg-[--Background] p-[3.2rem]">
       <div className="flex flex-col content-center items-center gap-[3.2rem]">
@@ -15,19 +15,19 @@ export default function Login() {
               <Image src={LogoIcon} alt="메인로고" width={210} height={38} />
             </Link>
             <div className="flex items-center justify-center gap-[0.8rem]">
-              <span className="text-[1.6rem]">회원이 아니신가요?</span>
-              <Link href="/signup" style={{ textDecoration: "none" }}>
+              <span className="text-[1.6rem]">이미 회원이신가요?</span>
+              <Link href="/login" style={{ textDecoration: "none" }}>
                 <p className="text-[1.6rem] font-[600] text-[--Primary]">
-                  회원 가입하기
+                  로그인하기
                 </p>
               </Link>
             </div>
           </div>
-          <LoginForm />
+          <SignUpForm />
         </div>
         <div className="flex w-[40rem] items-center justify-between rounded-[0.8rem] border border-solid border-[--Linkbrary-gray20] bg-[--Linkbrary-gray10] px-[2.4rem] py-[1.2rem]">
           <p className="text-[1.4rem] font-[--Linkbrary-gray100]">
-            소셜 로그인
+            다른 방식으로 가입하기
           </p>
           <div className="flex items-center justify-normal gap-[1.6rem]">
             <Link
@@ -55,3 +55,5 @@ export default function Login() {
     </div>
   );
 }
+
+export default SignUp;
